requirements-render.txt
# Minimal requirements for Render deployment
# Telegram Bot Framework
aiogram==3.4.1

# Database
asyncpg==0.29.0
SQLAlchemy==2.0.25

# HTTP Client
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# Image Processing (compatible with Python 3.13)
Pillow==10.4.0

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
